--- .babelrc.json
@@ -1,15 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": "> 2%, not dead, not ie > 0",
-        "modules": false
-      }
-    ]
-  ],
-  "env": {
-    "test": { "presets": [["@babel/preset-env"]] }
-  },
-  "plugins": ["@babel/plugin-proposal-class-properties"]
-}
