+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules/
-.cache/
 dist/
+.cache/
 yarn-error.log
