--- .yarn/build-state.yml
@@ -1,38 +1,0 @@
-# Warning: This file is automatically generated. Removing it is fine, but will
-# cause all your builds to become invalidated.
-
-# @web3-js/websocket@npm:1.0.30
-"5f6c45434ead0059bb56bac4dc8e497305fea224c080514c48af7e7b0470b6ca34bc8f76f9ea267f5441c343842e97bfc1ef1be83c8d0c69547ab4c22f2b1ad0":
-  506acc4dff09f3acc10cffaa455a664bc0fc462239ba4b18b6ca0aa820cc8175bbda3277e24ca1221037a6ccc044367473c1b8ec7416bcf32c9d2a7ceb97006a
-
-# core-js@npm:2.6.11
-"15178ded27ab674ae2054269453d809bdb1d00b98392a34947b5d43ea7a5811e5674c2fda7d48bb653b24a3506b0a8aa126bbac861bdeba93438ec6c7efb2d9d":
-  ce7b02ae66ea2eee25981a8f7a4b0389312ef9dc8980e169bfd5b1d0cd6de3ca2152729852f72b9130d42601bd677b3763642762088db45cd766fa3d9da5414f
-
-# eccrypto@npm:1.1.3
-"21829ddb6f07a1e80fcbf24c5ab9fb44e0617907185104fdf71888cc8435f1b06862940dae4c9144bbb47e18c88bbf5d43351e6416d5e71fada524d18ce15253":
-  2851eb8a4c792d7cc381a1a7b36e1f5ec8d3af534b4995aaef1134cb3bb193b7f80e7a71ec05e3b78ff8d016d9916fa328012cf1818a83efec92465eafd3fd4f
-
-# fsevents@patch:fsevents@npm%3A1.2.11#builtin<compat/fsevents>::version=1.2.11&hash=77dfe6
-"d3c7a0a32cad4399d02d3ba5cb50e4c30a13f4a107b288c7f254383861166489cb6e40523e99429d7acb0177b3cc98e3ed9395ffbc813495970067a35e5d3433":
-  b48f96e48f8dd0e19651f53c802bddeb0c9a8e1787220467bf6621a08b4c1762939399e13de09b61e8e91bdcab509378a4b8648265cbc706b774bfbabe5990b2
-
-# keccak@npm:1.4.0
-"53a72120ec800c22894226d3697f008f78974108cf87dc1c17c3b470e0d42317515eaaaaf656f88c8216711e339fe0b804b683aaa6629fe156d1a7a446bc114b":
-  d3035ef1613fc4d3bcfd516ec3f04b5fa1f0feaaca163395f8d4f4283ddcc37919c1c82012f833589dbe1f76c787860fa536a49b9ad9b2cf2d6e87702cd762d3
-
-# keccak@npm:2.1.0
-"b36e31bfb39dd9034c3ef94b6dc64c450cc0c2c230b3d7f87198ca6e35d21ceb81076a92d5aedfa3c23742ef76453a77b4fdae126a529bb64c88ebba185aa82b":
-  35e79bc03890366cb20d494e3395923b5906ab3700e47c5640d022791aadd6fac5026dca504fde8d4223729c23752c842ff96d43389d0470d124af4fff73a8ab
-
-# secp256k1@npm:3.7.1
-"b6fe8b03a52f04c3daa82be76a9d0127bf0b013b0905f8c22b562a4486cd8c4f90d52cb4bc495de09c26606f6f333cb6d203af95e9590ec7e77d9edcd4d56e58":
-  455275c3c0d4e6e45024659bff5fd04d6c2f89687354dbdc03fbdd8030e09ef7b426a4cefd93c356f86cdaab850aa24d4280c0f887414f347a42c3d6478fac42
-
-# secp256k1@npm:3.8.0
-"e2ef54940a998b00b8742766988bc2404b0f1e17460b5cd3403f069c7a4af0f6be68462f35aa847dde9afe17926c2505fed8d19c2c128ec9537f41d09dae2aae":
-  5b4d75f5db1d9fa7ea46dd55e47e7acf6f5d6ab98129405add2c7afbb00c8e48b32028538b5598cb288ac605ab1aba7ff0f645216e9697c5d896b79825c336b0
-
-# sha3@npm:1.2.6
-"b078b3c01436de07a5c6d2951e3c9b596fd9d0d23fad525740d7d3f63a17ce2c3ade97f2d42fc4381355001f9f01e09bd2322778b3278f30f9c40f3e797fe430":
-  b3cbd20fdb2786a8f00dd7b81760883139b7d9b454261f8283f28f84abc23cb80a1465a2d8566f43413068158a09aa8d56f370825c326c2c3723a95bf37a143f
