+++ package.json
@@ -3,48 +3,55 @@
   "version": "0.2.0",
   "author": "Aragon Association <legal@aragon.org>",
   "license": "MIT",
-  "main": "dist/cjs/index.js",
-  "module": "dist/index.js",
+  "main": "dist/index.js",
+  "module": "dist/token-amount.esm.js",
   "sideEffects": false,
+  "typings": "dist/index.d.ts",
   "files": [
     "dist"
   ],
-  "scripts": {
-    "build": "yarn rollup --config",
-    "dev": "yarn rollup --config --watch",
-    "lint": "yarn eslint ./src",
-    "test": "yarn lint && jest",
-    "prepare": "rm -rf ./dist && yarn build",
-    "prepublishOnly": "git push && git push --tags"
+  "engines": {
+    "node": ">=10"
   },
+  "scripts": {
+    "start": "tsdx watch --transpileOnly",
+    "build": "tsdx build",
+    "test": "tsdx test",
+    "lint": "tsdx lint src --max-warnings 0 && tsc",
+    "prepare": "tsdx build",
+    "prepublishOnly": "git push && git push --tags",
+    "size": "size-limit",
+    "analyze": "size-limit --why"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "yarn lint"
+    }
+  },
+  "size-limit": [
+    {
+      "path": "dist/token-amount.cjs.production.min.js",
+      "limit": "10 KB"
+    },
+    {
+      "path": "dist/token-amount.esm.js",
+      "limit": "10 KB"
+    }
+  ],
   "dependencies": {
-    "jsbi": "^3.1.1"
+    "jsbi": "^3.1.4"
   },
   "devDependencies": {
-    "@babel/cli": "^7.8.3",
-    "@babel/core": "^7.8.3",
-    "@babel/plugin-proposal-class-properties": "^7.10.1",
-    "@babel/preset-env": "^7.8.3",
-    "@babel/preset-react": "^7.8.3",
-    "@rollup/plugin-commonjs": "^11.0.1",
-    "@rollup/plugin-json": "^4.0.1",
-    "@rollup/plugin-node-resolve": "^7.0.0",
-    "babel-eslint": "^10.0.3",
-    "babel-jest": "^26.0.1",
-    "builtin-modules": "^3.1.0",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.9.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-plugin-import": "^2.20.0",
-    "eslint-plugin-jest": "^23.13.2",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-prettier": "^3.1.2",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
-    "jest": "^26.0.1",
-    "prettier": "^1.19.1",
-    "rollup": "^1.29.1",
-    "rollup-plugin-babel": "^4.3.3",
-    "rollup-plugin-terser": "^5.2.0"
-  }
+    "@size-limit/preset-small-lib": "^4.6.0",
+    "@typescript-eslint/eslint-plugin": "^4.3.0",
+    "@typescript-eslint/parser": "^4.3.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "husky": "^4.3.0",
+    "prettier": "^2.1.2",
+    "size-limit": "^4.6.0",
+    "tsdx": "^0.14.0",
+    "tslib": "^2.0.1",
+    "typescript": "^4.0.3"
+  },
+  "peerDependencies": {}
 }
