--- src/characters.js
@@ -1,1 +1,0 @@
-export const NO_BREAK_SPACE = '\u00a0'
