+++ src/characters.ts
@@ -1,0 +1,1 @@
+export const NO_BREAK_SPACE = '\u00a0'
