--- src/index.js
@@ -1,3 +1,0 @@
-import TokenAmount from './TokenAmount'
-
-export default TokenAmount
