+++ src/index.ts
@@ -1,0 +1,3 @@
+import TokenAmount from './TokenAmount'
+
+export default TokenAmount
